Renames time package to teatime for ease with conflicting with the standard time package. Adds method to a token to convert its expired after to a time. Adds Set method to a token repository
