+++ packages/teatime/clock_test.go
@@ -1,4 +1,4 @@
-package time
+package teatime
 
 import (
 	"testing"
