+++ packages/teatime/go.mod
@@ -1,0 +1,3 @@
+module theodo.red/creditcompanion/packages/teatime
+
+go 1.15
