--- packages/time/go.mod
@@ -1,3 +1,0 @@
-module theodo.red/creditcompanion/packages/time
-
-go 1.15
