+++ packages/tokens/models/token_repository.go
@@ -1,5 +1,6 @@
 package models
 
 type TokenRepository interface {
-	Get(id string) Token
+	Get(id string) (*Token, error)
+	Set(id string, token *Token) error
 }
